+++ src/main/java/com/paoloamosso/piggus/controller/DeadlineController.java
@@ -67,12 +67,20 @@ public class DeadlineController {
         return modelAndView;
     }
     @RequestMapping(value="/update-deadline", method=RequestMethod.POST)
-    public ModelAndView processDeadline(@ModelAttribute("deadline") Deadline deadline) {
+    public ModelAndView processDeadline(
+            @ModelAttribute("deadline") Deadline deadline,
+            @ModelAttribute("numberPeriod")int numberPeriod,
+            @ModelAttribute("period")int period,
+            @ModelAttribute("repeatedTimes")int repeatedTimes,
+            @ModelAttribute("multipleUpdate")int multipleUpdate) {
         ModelAndView modelAndView = new ModelAndView();
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         User user = userService.findUserByUsername(auth.getName());
         deadline.setUser(user);
-        deadlineService.saveDeadline(deadline);
+        if (multipleUpdate == 1) {
+            deadlineService.removeDeadline(deadline,1);
+        }
+        deadlineService.createRepeatedDeadlines(deadline,numberPeriod,period,repeatedTimes);
         modelAndView.setViewName("redirect:/home");
         return modelAndView;
     }
@@ -87,7 +95,7 @@ public class DeadlineController {
         if (user.getDeadlines().contains(deadline)) {
             deadline.setArchived(true);
             deadlineService.saveDeadline(deadline);
-            modelAndView.setViewName("redirect:/home");
+            modelAndView.setViewName("redirect:/deadline/list");
         } else {
             modelAndView.setViewName("redirect:/404");
         }
@@ -96,14 +104,15 @@ public class DeadlineController {
 
     // ==== REMOVE DEADLINE ====
     @RequestMapping(value="/remove-deadline", method=RequestMethod.GET)
-    public ModelAndView removeDeadline (@RequestParam(value="deadlineID",required = false) int deadlineID) {
+    public ModelAndView removeDeadline (@RequestParam(value="deadlineID",required = false) int deadlineID,
+                                        @RequestParam(value = "multipleRemove", required = false) int multipleRemove) {
         ModelAndView modelAndView = new ModelAndView();
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         User user = userService.findUserByUsername(auth.getName());
         Deadline deadline = deadlineService.getDeadline(deadlineID);
         if (user.getDeadlines().contains(deadline)) {
-            deadlineService.removeDeadline(deadline);
-            modelAndView.setViewName("redirect:/home");
+            deadlineService.removeDeadline(deadline,multipleRemove);
+            modelAndView.setViewName("redirect:/deadline/list");
         } else {
             modelAndView.setViewName("redirect:/404");
         }